requirements.txt
fastapi==0.128.0
uvicorn[standard]==0.40.0
SQLAlchemy==2.0.45
psycopg2-binary==2.9.11
python-dotenv==1.2.1
python-jose[cryptography]==3.5.0
bcrypt==5.0.0
python-multipart==0.0.21
pydantic[email]==2.12.5
pydantic-settings==2.12.0
